update processors to photcal restructure
