+++ mirar/pipelines/git/blocks.py
@@ -35,7 +35,8 @@ from mirar.processors.astromatic.sextractor.background_subtractor import (
 from mirar.processors.astromatic.swarp import Swarp
 from mirar.processors.cosmic_rays import LACosmicCleaner
 from mirar.processors.csvlog import CSVLog
-from mirar.processors.photcal import OutlierRejectionZPCalculator, PhotCalibrator
+from mirar.processors.photcal.photcal import PhotCalibrator
+from mirar.processors.photcal.zp_calculator import OutlierRejectionZPCalculator
 from mirar.processors.photometry import AperturePhotometry, PSFPhotometry
 from mirar.processors.reference import ProcessReference
 from mirar.processors.sources import ForcedPhotometryDetector
